Move db utils functions to utils_db file
